feat: missing api routes and swagger for all (#155)
* feat: get users

* feat: get comments

* feat: get users pagination

* refacto: getUrl url with url params

* feat: getPaginationSchemas for standardize all pagination routes

* refacto: type pagination schemas retunr

* fix: pagination schemas type

* feat: getPaginationSchemas for notifications

* feat: getPaginationSchemas for comments get

* feat: count on comments translation

* refacto: users routes with zod parse and swagger except PATCH User

* refacto: delete useless Downloads endpoint and rearrange endpoints in swagger

* refacto: images routes with zod parse and swagger

* feat: zod parse for user session and account route

* feat: swagger for get comments + get/patch/delete comment

* feat: swagger for comment routes

* feat: swagger for credential oauth

* feat: swagger for notifications

* feat: groupBy for notifications count

* refacto: comment is already deleted with 410 code

* fix: notifications stats

* feat: oauth credential post token

* fix: rearrange route and schemas of movie routes

* fix: less loop (#166)

* feat: get movies swagger with pagination with options

* feat: swagger for get movie and sub

* feat: swagger for like movie rotues

* fix: delete try catch for like and unlike

* feat: pagination schema in packages folder

* feat: swagger for movie comment get/post and rm try/catch from global comments functions

* feat: swagger for watchtimer and casting

* feat: swagger for authentification

* feat: swagger for users PATCH

---------

+++ apps/server/src/routes/global/like.global.ts
@@ -5,53 +5,49 @@ import { ContentfulStatusCode } from "hono/utils/http-status";
 export const likeParent = async (
   userId: string,
   parentId: string,
-  parentType: TParentType,
+  parentType: TParentType
 ): Promise<{ message: string; status: ContentfulStatusCode }> => {
-  try {
-    await prisma.like.create({
-      data: {
-        userId,
-        parentId,
-        parentType,
-      },
-    });
+  const like = await prisma.like.findUnique({
+    where: { userId_parentId: { userId, parentId } },
+  });
+  if (like) return { message: `${parentType} already liked`, status: 409 };
 
-    hypertubeLogger.info(`${parentType} liked successfully by user ${userId}`);
+  await prisma.like.create({
+    data: {
+      userId,
+      parentId,
+      parentType,
+    },
+  });
 
-    return { message: `${parentType} liked successfully`, status: 201 };
-  } catch (error) {
-    hypertubeLogger.error(`Error when liking ${parentType} : ${error}`);
-    return {
-      message: `Unexpected error when liking ${parentType}`,
-      status: 400,
-    };
-  }
+  hypertubeLogger.info(`${parentType} liked successfully by user ${userId}`);
+
+  return { message: `${parentType} liked successfully`, status: 201 };
 };
 
 export const unlikeParent = async (
   userId: string,
   parentId: string,
+  parentType: TParentType
 ): Promise<{ message: string; status: ContentfulStatusCode }> => {
-  try {
-    await prisma.like.delete({
-      where: {
-        userId_parentId: {
-          userId,
-          parentId,
-        },
-      },
-    });
+  const like = await prisma.like.findUnique({
+    where: { userId_parentId: { userId, parentId } },
+  });
+  if (!like) return { message: `${parentType} already unliked`, status: 409 };
 
-    hypertubeLogger.info(`${parentId} unliked successfully by ${userId}`);
-    return {
-      message: `${parentId} unliked successfully.`,
-      status: 200,
-    };
-  } catch (error) {
-    hypertubeLogger.error(`Error unliking comment: ${error}`);
-    return {
-      message: `Unexpected error when unliking ${parentId}`,
-      status: 400,
-    };
-  }
+  await prisma.like.delete({
+    where: {
+      userId_parentId: {
+        userId,
+        parentId,
+      },
+    },
+  });
+
+  hypertubeLogger.info(`${parentType} unliked successfully by ${userId}`);
+
+  return {
+    message: `${parentType} unliked successfully.`,
+    status: 200,
+  };
 };
